feat(core, dto): invert link between dto and core
