+++ packages/api/src/modules/user/services/activation/user.activation.service.ts
@@ -1,7 +1,14 @@
 import { ObjectId } from "mongodb";
 import * as RandToken from "rand-token";
-import { ResetPasswordErrorCodes, TokenValidationDtoResponse, TokenValidationType, UserDto } from "dto";
-import { BadRequestError, InternalServerError, NotFoundError, ResetTokenNotFoundError, UserNotFoundError } from "core";
+import { TokenValidationDtoResponse, TokenValidationType, UserDto } from "dto";
+import {
+    BadRequestError,
+    InternalServerError,
+    NotFoundError,
+    ResetPasswordErrorCodes,
+    ResetTokenNotFoundError,
+    UserNotFoundError,
+} from "core";
 import userAdapter from "../../../../adapters/outputs/db/user/user.adapter";
 import { JWT_EXPIRES_TIME } from "../../../../configurations/jwt.conf";
 import UserReset from "../../entities/UserReset";
