+++ packages/core/constants/common.ts
@@ -1,4 +1,1 @@
-import type { NOT_APPLICABLE } from "dto";
-
-// TODO: move this in core
-export const NOT_APPLICABLE_VALUE: NOT_APPLICABLE = "N/A";
+export const NOT_APPLICABLE_VALUE = "N/A";
