+++ packages/core/enums/LoginErrorCodes.ts
@@ -1,0 +1,6 @@
+export enum LoginErrorCodes {
+    EMAIL_OR_PASSWORD_NOT_MATCH = 1,
+    USER_NOT_ACTIVE = 2,
+    INTERNAL_ERROR = 3,
+    PASSWORD_UNSET = 4,
+}
