+++ packages/core/enums/ResetPasswordErrorCodes.ts
@@ -1,0 +1,8 @@
+export enum ResetPasswordErrorCodes {
+    RESET_TOKEN_NOT_FOUND = 1,
+    RESET_TOKEN_EXPIRED = 2,
+    USER_NOT_FOUND = 3,
+    PASSWORD_FORMAT_INVALID = 4,
+    INTERNAL_ERROR = 5,
+    PROCONNECT_NO_RESET = 6,
+}
