+++ packages/core/enums/SearchErrorCodes.ts
@@ -1,0 +1,4 @@
+export enum SearchErrorCodes {
+    ID_NOT_ASSO = 0,
+    MULTIPLE_ASSOS = 1,
+}
