+++ packages/core/errors/LoginError.ts
@@ -1,8 +1,8 @@
-import { LoginDtoErrorCodes } from "dto";
+import { LoginErrorCodes } from "../enums/LoginErrorCodes";
 import { UnauthorizedError } from "./httpErrors";
 
 export class LoginError extends UnauthorizedError {
     constructor() {
-        super("Invalid credentials", LoginDtoErrorCodes.EMAIL_OR_PASSWORD_NOT_MATCH);
+        super("Invalid credentials", LoginErrorCodes.EMAIL_OR_PASSWORD_NOT_MATCH);
     }
 }
