+++ packages/core/errors/MultipleAssociationsError.ts
@@ -1,11 +1,11 @@
-import { SearchCodeError } from "dto";
+import { SearchErrorCodes } from "../enums/SearchErrorCodes";
 import { BadRequestError } from "./httpErrors";
 
 export class MultipleAssociationsError extends BadRequestError {
     constructor() {
         super(
             "Multiple associations found with this identifier, please use a more specific one",
-            SearchCodeError.MULTIPLE_ASSOS,
+            SearchErrorCodes.MULTIPLE_ASSOS,
         );
     }
 }
