+++ packages/core/errors/NotAssociationError.ts
@@ -1,8 +1,8 @@
-import { SearchCodeError } from "dto";
+import { SearchErrorCodes } from "../enums/SearchErrorCodes";
 import { UnprocessableError } from "./httpErrors/UnprocessableError";
 
 export class NotAssociationError extends UnprocessableError {
     constructor() {
-        super("Votre recherche pointe vers une entité qui n'est pas une association", SearchCodeError.ID_NOT_ASSO);
+        super("Votre recherche pointe vers une entité qui n'est pas une association", SearchErrorCodes.ID_NOT_ASSO);
     }
 }
