+++ packages/core/errors/index.ts
@@ -1,6 +1,5 @@
 export * from "./httpErrors";
 export * from "./AssociationIdentifierError";
-export * from "./GrantError";
 export * from "./MultipleAssociationsError";
 export * from "./LoginError";
 export * from "./NotAssociationError";
