+++ packages/dto/package.json
@@ -43,6 +43,7 @@
     },
     "dependencies": {
         "@types/node": "24.10.1",
-        "mongodb": "6.16.0"
+        "mongodb": "6.16.0",
+        "core": "workspace:0.84.0"
     }
 }
