+++ packages/dto/src/auth/LoginDtoResponse.ts
@@ -1,11 +1,5 @@
 import { UserWithJWTDto } from "../user/UserDto";
-
-export enum LoginDtoErrorCodes {
-    EMAIL_OR_PASSWORD_NOT_MATCH = 1,
-    USER_NOT_ACTIVE = 2,
-    INTERNAL_ERROR = 3,
-    PASSWORD_UNSET = 4,
-}
+import { LoginErrorCodes } from "core";
 
 export interface LoginDtoPositiveResponse {
     user: UserWithJWTDto;
@@ -13,7 +7,7 @@ export interface LoginDtoPositiveResponse {
 
 export interface LoginDtoNegativeResponse {
     message: string;
-    code: LoginDtoErrorCodes;
+    code: LoginErrorCodes;
 }
 
 export type LoginDtoResponse = LoginDtoPositiveResponse | LoginDtoNegativeResponse;
