--- packages/dto/src/search/SearchCodeError.ts
@@ -1,4 +1,0 @@
-export enum SearchCodeError {
-    ID_NOT_ASSO,
-    MULTIPLE_ASSOS,
-}
