+++ packages/dto/src/search/index.ts
@@ -1,2 +1,1 @@
 export * from "./AssociationNameDto";
-export * from "./SearchCodeError";
